// backend/src/dto/auth.dto.ts
import { z } from 'zod';

export const RegisterSchema = z.object({
    email: z.email(),
    password: z.string().min(6),
    firstname: z.string().min(1).optional(),
    lastname: z.string().min(1).optional(),
});

export type RegisterDto = z.infer<typeof RegisterSchema>;

export const LoginSchema = z.object({
    email: z.string().email(),
    password: z.string().min(1),
});

export type LoginDto = z.infer<typeof LoginSchema>;